move event to paper
