+++ paper/src/main/java/com/denizenscript/denizen/paper/events/PlayerLinksSendScriptEvent.java
@@ -1,11 +1,12 @@
-package com.denizenscript.denizen.events.player;
+package com.denizenscript.denizen.paper.events;
 
 import com.denizenscript.denizen.events.BukkitScriptEvent;
 import com.denizenscript.denizen.utilities.Utilities;
 import com.denizenscript.denizen.utilities.implementation.BukkitScriptEntryData;
 import com.denizenscript.denizencore.objects.core.ListTag;
 import com.denizenscript.denizencore.scripts.ScriptEntryData;
-import com.denizenscript.denizencore.utilities.debugging.Debug;
+import io.papermc.paper.connection.PlayerGameConnection;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerLinksSendEvent;
@@ -18,16 +19,19 @@ public class PlayerLinksSendScriptEvent extends BukkitScriptEvent implements Lis
     //
     // @Group Player
     //
-    // @Triggers when a player opens server links.
+    // @Plugin Paper
+    //
+    // @Triggers when the list of links is sent to the player
     //
     // @Determine
     // "LINKS:<ListTag(MapTag)>" to change what links should player see. Each map needs to have "uri" key which represents URI address of this link and either display or a type. Valid types are listed at <@link url https://hub.spigotmc.org/javadocs/spigot/org/bukkit/ServerLinks.Type.html>
     //
-    // @Player Always.
+    // @Player when the event part of player game connection
     //
     // -->
 
     public PlayerLinksSendEvent event;
+    public Player player;
 
     public PlayerLinksSendScriptEvent() {
         registerCouldMatcher("player links send");
@@ -38,11 +42,12 @@ public class PlayerLinksSendScriptEvent extends BukkitScriptEvent implements Lis
 
     @Override
     public ScriptEntryData getScriptEntryData() {
-        return new BukkitScriptEntryData(event.getPlayer());
+        return new BukkitScriptEntryData(player);
     }
 
     @EventHandler
     public void onPlayerLinksSend(PlayerLinksSendEvent event) {
+        player = event.getConnection() instanceof PlayerGameConnection connection ? connection.getPlayer() : null;
         this.event = event;
         fire(event);
     }
